Move LoginForm to login and refactor form into separate components
